+++ packages/logger/src/index.ts
@@ -1,5 +1,5 @@
 import { cyan, red, magenta, bold } from 'colorette'
-import statusEmoji from 'http-status-emojis';
+import statusEmoji from 'http-status-emojis'
 import dayjs from 'dayjs'
 import { IncomingMessage as Request, ServerResponse as Response, METHODS } from 'http'
 
@@ -7,20 +7,17 @@ export interface LoggerOptions {
   methods?: string[]
   output?: {
     color: boolean
-    callback: (string) => void
+    callback: (string: string) => void
   }
   timestamp?: boolean | { format?: string }
-  badges?: {
-    emoji: boolean
-    captions: boolean
-  }
+  emoji?: boolean
 }
 
 export const logger = (options: LoggerOptions = {}) => {
   const methods = options.methods ?? METHODS
   const timestamp = options.timestamp ?? false
   const output = options.output ?? { callback: console.log, color: true }
-  const badge = options.badges ?? { emoji: false, captions: false };
+  const emojiEnabled = options.emoji
 
   return (req: Request, res: Response, next?: () => void) => {
     res.on('finish', () => {
@@ -42,30 +39,27 @@ export const logger = (options: LoggerOptions = {}) => {
           }
         }
 
-        let badges = '';
-        if (badge.emoji) {
-          badges = statusEmoji[s];
-        } 
+        const emoji = emojiEnabled ? statusEmoji[s] : ''
 
         if (!output.color) {
-          const m = `${badges} ${time}${method} ${status} ${msg} ${url}`;
+          const m = `${emoji} ${time}${method} ${status} ${msg} ${url}`
           output.callback(m)
         } else {
           switch (s[0]) {
             case '2':
               status = cyan(bold(s))
               msg = cyan(msg)
-              output.callback(`${badges} ${time}${method} ${status} ${msg} ${url}`)
+              output.callback(`${emoji} ${time}${method} ${status} ${msg} ${url}`)
               break
             case '4':
               status = red(bold(s))
               msg = red(msg)
-              output.callback(`${badges} ${time}${method} ${status} ${msg} ${url}`)
+              output.callback(`${emoji} ${time}${method} ${status} ${msg} ${url}`)
               break
             case '5':
               status = magenta(bold(s))
               msg = magenta(msg)
-              output.callback(`${badges} ${time}${method} ${status} ${msg} ${url}`)
+              output.callback(`${emoji} ${time}${method} ${status} ${msg} ${url}`)
               break
           }
         }
@@ -75,4 +69,3 @@ export const logger = (options: LoggerOptions = {}) => {
     next?.()
   }
 }
-